fix: add class and method about comment and like

+++ src/main/java/com/salle/server/repository/ProductImageRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface ProductImageRepository extends JpaRepository<ProductImage, Integer> {
+public interface ProductImageRepository extends JpaRepository<ProductImage, Long> {
 
     List<String> findByProduct(Product product);
 }
